made extra component of timer and aligned options button prorperly in pomoview
